refactor: 출력 관련 로직 WebOutput으로 통일
